Start rewrite of PFEGammaAlgo from the ground up. Much of the old code is sitting around, simply for reference. The 'unwrapSuperCluster' method isn't implemented yet, but we know what it does. The first two basic steps of making basic proto-objects out of SCs and seeded GSF tracks are there.
This just compiles and will not yet actually run this code.
To see additions for the start of the rewrite look in PFEGammaAlgoNew and search for "rewritten" and then
you can follow the function names to see the logic.

PFEGammaProducerNew needs to have the refined supercluster part turned into a function that eats a PFCandidate
and returns a reco::SuperCluster, harvesting the ECAL and PS components of the refined EGamma PFCandidate.